% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_tryptase.R
\name{count_tryptase}
\alias{count_tryptase}
\title{Count Tryptase Alleles in a .sam File}
\usage{
count_tryptase(sam_file)
}
\arguments{
\item{sam_file}{path to the temp.sam file generated by the pre-processing}
}
\description{
Count Tryptase Alleles in a .sam File
}
