% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_tryptase.R
\name{preprocess}
\alias{preprocess}
\title{Run preprocessing script}
\usage{
preprocess(cram_file)
}
\arguments{
\item{cram_file}{file path of .cram file to be processed}
}
\value{
produces a temp.sam file
}
\description{
Run preprocessing script
}
