% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrypLik.R
\name{select_population}
\alias{select_population}
\title{Select a genetic-ancestry population}
\usage{
select_population(pop = NULL)
}
\arguments{
\item{pop}{population to use (0=EUR, 1=AFR, 2=EAS, 3=SAS) as either a number
or character string}
}
\description{
Select a genetic-ancestry population
}
