% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrypLik-package.R
\docType{package}
\name{TrypLik-package}
\alias{TrypLik}
\alias{TrypLik-package}
\title{TrypLik: Calculates Tryptase Variant Likelihoods}
\description{
Calculates tryptase variant likelihoods based on counts.
}
\author{
\strong{Maintainer}: Jonathan Carroll \email{jonathan@hibio.com} (\href{https://orcid.org/0000-0002-1404-5264}{ORCID})

Authors:
\itemize{
  \item Jeff Wall \email{jeffwall.genetics@gmail.com}
}

Other contributors:
\itemize{
  \item Human Immunology Biosciences (HI·Bio) [copyright holder, funder]
}

}
\keyword{internal}
