% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{tryptase_1KG}
\alias{tryptase_1KG}
\title{Retrieve the 1000 Genomes data, processed for tryptase allele counts}
\usage{
tryptase_1KG()
}
\value{
A \code{data.frame} of 2A/D, 2B/D, and the inferred genotype group for the
503 relevant samples
}
\description{
See the background vignette of \code{TrypLik}
for additional details.
}
\examples{
head(tryptase_1KG())
}
