% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_tryptase.R
\name{preprocess}
\alias{preprocess}
\title{Run preprocessing script}
\usage{
preprocess(
  cram_file,
  output_dir = dirname(cram_file),
  consensus_dir = system.file(package = "TrypLik"),
  nthreads = 8
)
}
\arguments{
\item{cram_file}{file path of .cram file to be processed}
}
\value{
produces a temp.sam file
}
\description{
Run preprocessing script
}
\examples{
tryCatch(preprocess("missing_file.cram"), error = function(e) e)

\dontrun{
preprocess("HG00100.final.cram")
}
}
