% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrypLik.R
\name{TrypLik}
\alias{TrypLik}
\title{Calculate Tryptase Variant Likelihoods}
\usage{
TrypLik(wt, fs, areads, breads, dreads, pop = 0)
}
\arguments{
\item{wt}{wildtype}

\item{fs}{frameshift}

\item{areads}{alpha counts}

\item{breads}{beta counts}

\item{dreads}{delta counts}

\item{pop}{population to use (0=EUR, 1=AFR, 2=EAS, 3=SAS) as either a number
or character string}
}
\value{
a \code{data.frame} containing probabilities of the relevant likelihoods
}
\description{
Calculate Tryptase Variant Likelihoods
}
\examples{
# as individual arguments
TrypLik(26, 0, 347, 316, 304, pop = 0)

# as a `counts` vector
TrypLik(c(26, 0, 347, 316, 304), pop = 0)

# as the output of `preprocess`
TrypLik(c(24, 0, 22, 0, 26, 0, 347, 316, 304), pop = 0)

}
